[BE] MOVE: dto 폴더에 event, request, response 세부 폴더 추가 #85
