+++ src/backend/main-server/main/src/main/java/com/kickzo/main/dto/request/RoomUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.kickzo.main.dto;
+package com.kickzo.main.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
