+++ src/backend/main-server/main/src/main/java/com/kickzo/main/dto/response/RoomResponseDto.java
@@ -1,18 +1,24 @@
-package com.kickzo.main.dto;
+package com.kickzo.main.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
-@Data
+@Getter
+@Setter
 @AllArgsConstructor
+@NoArgsConstructor
 @Builder
-public class RoomInfoDto {
+public class RoomResponseDto {
 	private Long id;
 	private String code;
 	private String title;
 	private String description;
-	private int userCount;
 	private String creator;
 	private String profileImageUrl;
+	private int userCount;
+	private String playlistUrl;
 }
+
