+++ src/backend/main-server/main/src/main/java/com/kickzo/main/service/MainPageService.java
@@ -13,9 +13,9 @@ import org.springframework.transaction.annotation.Transactional;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.kickzo.main.dto.CreateRoomRequestDto;
-import com.kickzo.main.dto.CreateRoomResponseDto;
-import com.kickzo.main.dto.RoomResponseDto;
+import com.kickzo.main.dto.request.CreateRoomRequestDto;
+import com.kickzo.main.dto.response.CreateRoomResponseDto;
+import com.kickzo.main.dto.response.RoomResponseDto;
 import com.kickzo.main.entity.Room;
 import com.kickzo.main.entity.RoomUser;
 import com.kickzo.main.entity.RoomUserId;
